Refactor crypto functions and update imports
- Moved access key generation functions to a new `crypto.ts` file.
- Replaced `crypto.randomUUID()` with `uuid.v4()` for generating UUIDs.
- Updated imports across various files to reflect the new structure and removed unused imports.
