+++ src/app/api/v1/products/[account_id]/[repository_id]/members/route.ts
@@ -8,8 +8,8 @@ import {
 } from "@/types";
 import { StatusCodes } from "http-status-codes";
 import { isAuthorized } from "@/lib/api/authz";
-import * as crypto from "crypto";
 import { getApiSession } from "@/lib/api/utils";
+import { v4 as uuidv4 } from "uuid";
 import {
   accountsTable,
   membershipsTable,
@@ -94,7 +94,7 @@ export async function POST(
     }
     const membership: Membership = {
       ...membershipInvitation,
-      membership_id: crypto.randomUUID(),
+      membership_id: uuidv4(),
       membership_account_id: product.account_id,
       repository_id: product.product_id,
       state: MembershipState.Invited,
