+++ src/lib/actions/account.ts
@@ -20,7 +20,7 @@ import { getPageSession } from "../api/utils";
 import { accountsTable, membershipsTable } from "../clients";
 import { FormState } from "@/components/core/DynamicForm";
 import { redirect } from "next/navigation";
-import { randomUUID } from "crypto";
+import { v4 as uuidv4 } from "uuid";
 
 /**
  * Server action to record email verification timestamp
@@ -155,7 +155,7 @@ export async function createAccount(
       role: MembershipRole.Owners,
       state: MembershipState.Member,
       state_changed: new Date().toISOString(),
-      membership_id: randomUUID(),
+      membership_id: uuidv4(),
       membership_account_id: account.account_id,
     });
   }
